FileSystemCache Fix for droping a object that will probably be loaded async

+++ Spezifikation/src/de/joglearth/source/caching/FileSystemCache.java
@@ -8,7 +8,9 @@ import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.attribute.BasicFileAttributes;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -28,7 +30,10 @@ public class FileSystemCache<Key> implements Cache<Key, byte[]> {
     private Path basePath;
     private PathTranslator<Key> pathTranslator;
     private Set<Key> keySet;
+    private Set<Key> lockedFileSet;
     private ExecutorService executorService;
+    private Set<Key> filesForDroping;
+    private Map<Key, Set<SourceListener<Key, byte[]>>> registredListeners;
 
 
     /**
@@ -45,6 +50,9 @@ public class FileSystemCache<Key> implements Cache<Key, byte[]> {
         this.basePath = Paths.get(folder);
         this.pathTranslator = p;
         this.keySet = new HashSet<Key>();
+        this.lockedFileSet = new HashSet<Key>();
+        this.filesForDroping = new HashSet<Key>();
+        this.registredListeners = new HashMap<Key, Set<SourceListener<Key, byte[]>>>();
         this.executorService = Executors.newFixedThreadPool(5);
     }
 
@@ -53,12 +61,14 @@ public class FileSystemCache<Key> implements Cache<Key, byte[]> {
             SourceListener<Key, byte[]> sender) {
         SourceResponseType responseType;
         if (keySet.contains(key)) {
+            registerListener(key, sender);
+            lockedFileSet.add(key);
             responseType = SourceResponseType.ASYNCHRONOUS;
-            executorService.execute(new FileLoaderRunnable(key, sender));
+            executorService.execute(new FileLoaderRunnable(key));
         } else {
             responseType = SourceResponseType.MISSING;
         }
-        
+
         System.err.println("FileSystemCache: requesting " + key + ": " + responseType.toString());
         return new SourceResponse<byte[]>(responseType, null);
     }
@@ -83,8 +93,12 @@ public class FileSystemCache<Key> implements Cache<Key, byte[]> {
 
     @Override
     public synchronized void dropObject(Key k) {
-        System.err.println("FileSystemCache: dropping key " + k);
         keySet.remove(k);
+        if (lockedFileSet.contains(k)) {
+            filesForDroping.add(k);
+            return;
+        }
+        System.err.println("FileSystemCache: dropping key " + k);
         try {
             Files.deleteIfExists(pathFromKey(k));
         } catch (IOException e) {
@@ -105,8 +119,14 @@ public class FileSystemCache<Key> implements Cache<Key, byte[]> {
 
     @Override
     public synchronized void dropAll() {
-        System.err.println("FileSystemCache: dropping all objects");
         keySet.clear();
+        if (lockedFileSet.size() > 0) {
+            for (Key key : keySet) {
+                filesForDroping.add(key);
+            }
+            return;
+        }
+        System.err.println("FileSystemCache: dropping all objects");
         try {
             Files.walkFileTree(basePath, new DirectoryCleaner());
         } catch (IOException e) {
@@ -120,6 +140,24 @@ public class FileSystemCache<Key> implements Cache<Key, byte[]> {
         return filePath;
     }
 
+    private void registerListener(Key k, SourceListener<Key, byte[]> sender) {
+        Set<SourceListener<Key, byte[]>> listenerSet = registredListeners.get(k);
+        if (listenerSet == null) {
+            listenerSet = new HashSet<SourceListener<Key, byte[]>>();
+            registredListeners.put(k, listenerSet);
+        }
+        listenerSet.add(sender);
+    }
+
+    private void callListener(Key k, byte[] data) {
+        Set<SourceListener<Key, byte[]>> listenerSet = registredListeners.get(k);
+        if (listenerSet == null)
+            return;
+        for (SourceListener<Key, byte[]> sourceListener : listenerSet) {
+            sourceListener.requestCompleted(k, data);
+        }
+        listenerSet.clear();
+    }
 
     /**
      * Deletes all files inside a Path and its subpaths.
@@ -203,24 +241,37 @@ public class FileSystemCache<Key> implements Cache<Key, byte[]> {
             return FileVisitResult.TERMINATE;
         }
     }
-    
+
     private class FileLoaderRunnable implements Runnable {
+
         private Key key;
-        private SourceListener<Key, byte[]> listener;
-        public FileLoaderRunnable(Key k, SourceListener<Key, byte[]> l) {
+
+
+        public FileLoaderRunnable(Key k) {
             this.key = k;
-            this.listener = l;
         }
+
         @Override
         public void run() {
             try {
                 byte[] content = Files.readAllBytes(pathFromKey(key));
-                listener.requestCompleted(key, content);
+                callListener(key, content);
             } catch (IOException e) {
-                listener.requestCompleted(key, null);
+                callListener(key, null);
                 e.printStackTrace();
             }
+            lockedFileSet.remove(key);
+            synchronized (FileSystemCache.this) {
+                for (Key k : filesForDroping) {
+                    System.err.println("FileSystemCache: dropping key " + k);
+                    try {
+                        Files.deleteIfExists(pathFromKey(k));
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+            }
         }
-        
+
     }
 }
